started 2023 18

+++ rust/template.rs
@@ -5,8 +5,8 @@ use aoc_rust::stdin_lines;
 fn main() {
     let xs: Vec<XX> = stdin_lines().map(|s| parse(s)).collect();
 
-    let part_a: i32 = 0;
-    let part_b: i32 = 0;
+    let part_a: usize = 0;
+    let part_b: usize = 0;
 
     println!("{}", part_a);
     println!("{}", part_b);
